third_party/blink/web_tests/platform/generic/fast/url/trivial-segments-expected.txt

Test basic features of URL segmentation

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS segments('http://example.com/') is '["http:","example.com","","/","",""]'
PASS successfullyParsed is true

TEST COMPLETE
